int to long long

+++ shelkovnikov.dmitrii/common/inputDict.cpp
@@ -1,10 +1,10 @@
 #include "inputDict.h"
 #include <istream>
-dimkashelk::TwoThreeTree< int, std::string, std::less< >> dimkashelk::inputDict(std::istream &in)
+dimkashelk::TwoThreeTree< long long, std::string, std::less< >> dimkashelk::inputDict(std::istream &in)
 {
-  using dict_type = dimkashelk::TwoThreeTree< int, std::string, std::less< > >;
+  using dict_type = dimkashelk::TwoThreeTree< long long, std::string, std::less< > >;
   dict_type data;
-  int key = 0;
+  long long key = 0;
   std::string value;
   while (in)
   {
